edf-catalogacion-qrt/requirements.txt
alembic==1.13.2
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
dnspython==2.6.1
email-validator==1.1.3
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==20.1.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mongoengine==0.24.1
packaging==24.1
pdfkit==1.0.0
Pillow==8.4.0
pymongo==3.11.0
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
setuptools==70.2.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
WTForms==3.1.2
